+++ src/main/kotlin/com/knu/mockin/model/dto/request/quotations/basic/mock/CurrentPriceRequestParameterDto.kt
@@ -1,4 +1,4 @@
-package com.knu.mockin.model.dto.request.basic
+package com.knu.mockin.model.dto.request.quotations.basic.mock
 
 import com.fasterxml.jackson.annotation.JsonProperty
